Move dialog implementations and specs to their proper files
