+++ lib/alexandria/ui/alert_dialog.rb
@@ -4,7 +4,6 @@
 #
 # See the file README.md for authorship and licensing information.
 
-# HIG compliant error dialog boxes
 module Alexandria
   module UI
     class AlertDialog < SimpleDelegator
@@ -43,20 +42,6 @@ module Alexandria
           label.markup = message.strip
           vbox.pack_start(label)
         end
-      end
-    end
-
-    class ErrorDialog < AlertDialog
-      def initialize(parent, title, message = nil)
-        super(parent, title, Gtk::Stock::DIALOG_ERROR,
-              [[Gtk::Stock::OK, :ok]], message)
-        # FIXME: Should accept just :ok
-        self.default_response = Gtk::ResponseType::OK
-      end
-
-      def display
-        show_all && run
-        destroy
       end
     end
   end
