+++ lib/alexandria/ui/import_dialog.rb
@@ -4,6 +4,9 @@
 #
 # See the file README.md for authorship and licensing information.
 
+require 'alexandria/ui/error_dialog'
+require 'alexandria/ui/skip_entry_dialog'
+
 class Alexandria::ImportFilter
   def to_filefilter
     filefilter = Gtk::FileFilter.new
@@ -15,27 +18,6 @@ end
 
 module Alexandria
   module UI
-    class SkipEntryDialog < AlertDialog
-      include GetText
-      GetText.bindtextdomain(Alexandria::TEXTDOMAIN, charset: 'UTF-8')
-
-      def initialize(parent, message)
-        super(parent, _('Error while importing'),
-              Gtk::Stock::DIALOG_QUESTION,
-              [[Gtk::Stock::CANCEL, Gtk::ResponseType::CANCEL],
-               [_('_Continue'), Gtk::ResponseType::OK]],
-              message)
-        puts "Opened SkipEntryDialog #{inspect}" if $DEBUG
-        self.default_response = Gtk::ResponseType::CANCEL
-      end
-
-      def continue?
-        show_all && (@response = run)
-        destroy
-        @response == Gtk::ResponseType::OK
-      end
-    end
-
     class ImportDialog < SimpleDelegator
       include GetText
       include Logging
