+++ lib/alexandria/ui/keep_bad_isbn_dialog.rb
@@ -4,6 +4,8 @@
 #
 # See the file README.md for authorship and licensing information.
 
+require 'alexandria/ui/alert_dialog'
+
 module Alexandria
   module UI
     class KeepBadISBNDialog < AlertDialog
