+++ lib/alexandria/ui/new_book_dialog_manual.rb
@@ -5,6 +5,7 @@
 # See the file README.md for authorship and licensing information.
 
 require 'tmpdir'
+require 'alexandria/ui/error_dialog'
 
 module Alexandria
   module UI
