+++ lib/alexandria/ui/preferences_dialog.rb
@@ -4,205 +4,14 @@
 #
 # See the file README.md for authorship and licensing information.
 
-require 'alexandria/ui/builder_base'
 require 'alexandria/scanners/cue_cat'
 require 'alexandria/scanners/keyboard'
-
-class Gtk::Entry
-  attr_writer :mandatory
-  def mandatory?
-    @mandatory
-  end
-end
+require 'alexandria/ui/builder_base'
+require 'alexandria/ui/provider_preferences_dialog'
+require 'alexandria/ui/new_provider_dialog'
 
 module Alexandria
   module UI
-    class ProviderPreferencesBaseDialog < SimpleDelegator
-      def initialize(title:, parent:, flags:, buttons:)
-        dialog = Gtk::Dialog.new(title: title, parent: parent, flags: flags, buttons: buttons)
-        super(dialog)
-
-        self.resizable = false
-        child.border_width = 12
-
-        @controls = []
-      end
-
-      def fill_table(table, provider)
-        i = table.n_rows
-        table.resize(table.n_rows + provider.prefs.length,
-                     table.n_columns)
-        table.border_width = 12
-        table.row_spacings = 6
-        table.column_spacings = 12
-
-        @controls.clear
-
-        provider.prefs.read.each do |variable|
-          if variable.name == 'piggyback'
-            next
-            # ULTRA-HACK!! for bug #13302
-            # not displaying the visual choice, as its usually unnecessary
-            # Either way, this is confusing to the user: FIX
-            #    -   Cathal Mc Ginley 2008-02-18
-          end
-
-          if variable.name == 'enabled'
-            # also don't display Enabled/Disabled
-            next
-          end
-
-          label = Gtk::Label.new('_' + variable.description + ':')
-          label.use_underline = true
-          label.xalign = 0
-          table.attach_defaults(label, 0, 1, i, i + 1)
-
-          if variable.possible_values.nil?
-            entry = Gtk::Entry.new
-            entry.text = variable.value.to_s
-            entry.mandatory = variable.mandatory?
-          else
-            entry = Gtk::ComboBoxText.new
-            variable.possible_values.each do |value|
-              entry.append_text(value.to_s)
-            end
-            index = variable.possible_values.index(variable.value)
-            entry.active = index
-          end
-          label.mnemonic_widget = entry
-
-          @controls << [variable, entry]
-
-          table.attach_defaults(entry, 1, 2, i, i + 1)
-          i += 1
-        end
-        table
-      end
-
-      def sync_variables
-        @controls.each do |variable, entry|
-          variable.new_value = case entry
-                               when Gtk::ComboBox
-                                 variable.possible_values[entry.active]
-                               when Gtk::Entry
-                                 entry.text
-                               end
-        end
-      end
-    end
-
-    class ProviderPreferencesDialog < ProviderPreferencesBaseDialog
-      include GetText
-      GetText.bindtextdomain(Alexandria::TEXTDOMAIN, charset: 'UTF-8')
-
-      def initialize(parent, provider)
-        super(title: _('Preferences for %s') % provider.fullname,
-              parent: parent,
-              flags: :modal,
-              buttons: [[Gtk::Stock::CLOSE, :close]])
-
-        table = Gtk::Table.new(0, 0)
-        fill_table(table, provider)
-        child.pack_start(table)
-
-        signal_connect('destroy') { sync_variables }
-      end
-
-      def acquire
-        show_all
-        run
-        destroy
-      end
-    end
-
-    class NewProviderDialog < ProviderPreferencesBaseDialog
-      include GetText
-      GetText.bindtextdomain(Alexandria::TEXTDOMAIN, charset: 'UTF-8')
-
-      def initialize(parent)
-        super(title: _('New Provider'),
-              parent: parent,
-              flags: :modal,
-              buttons: [[Gtk::Stock::CANCEL, Gtk::ResponseType::CANCEL]])
-        @add_button = add_button(Gtk::Stock::ADD,
-                                 Gtk::ResponseType::ACCEPT)
-
-        instances = BookProviders.abstract_classes.map(&:new)
-        @selected_instance = nil
-
-        @table = Gtk::Table.new(2, 2)
-        child.pack_start(@table)
-
-        # Name.
-
-        label_name = Gtk::Label.new(_('_Name:'))
-        label_name.use_underline = true
-        label_name.xalign = 0
-        @table.attach_defaults(label_name, 0, 1, 0, 1)
-
-        entry_name = Gtk::Entry.new
-        entry_name.mandatory = true
-        label_name.mnemonic_widget = entry_name
-        @table.attach_defaults(entry_name, 1, 2, 0, 1)
-
-        # Type.
-
-        label_type = Gtk::Label.new(_('_Type:'))
-        label_type.use_underline = true
-        label_type.xalign = 0
-        @table.attach_defaults(label_type, 0, 1, 1, 2)
-
-        combo_type = Gtk::ComboBoxText.new
-        instances.each do |instance|
-          combo_type.append_text(instance.name)
-        end
-        combo_type.signal_connect('changed') do |cb|
-          @selected_instance = instances[cb.active]
-          fill_table(@table, @selected_instance)
-          sensitize
-          # FIXME: this should be re-written once we have multiple
-          # abstract providers.
-        end
-        combo_type.active = 0
-        label_type.mnemonic_widget = combo_type
-        @table.attach_defaults(combo_type, 1, 2, 1, 2)
-      end
-
-      def acquire
-        show_all
-        if run == Gtk::ResponseType::ACCEPT
-          @selected_instance.reinitialize(entry_name.text)
-          sync_variables
-        else
-          @selected_instance = nil
-        end
-        destroy
-      end
-
-      def instance
-        @selected_instance
-      end
-
-      private
-
-      def sensitize
-        entries = @table.children.select { |x| x.is_a?(Gtk::Entry) }
-        entries.each do |entry|
-          entry.signal_connect('changed') do
-            sensitive = true
-            entries.each do |entry2|
-              if entry2.mandatory?
-                sensitive = !entry2.text.strip.empty?
-                break unless sensitive
-              end
-            end
-            @add_button.sensitive = sensitive
-          end
-        end
-        @add_button.sensitive = false
-      end
-    end
-
     class PreferencesDialog < BuilderBase
       include Alexandria::Logging
       include GetText
