+++ lib/alexandria/ui/sidepane_manager.rb
@@ -4,6 +4,8 @@
 #
 # See the file README.md for authorship and licensing information.
 
+require 'alexandria/ui/error_dialog'
+
 module Alexandria
   module UI
     class SidePaneManager
