+++ spec/alexandria/ui/book_properties_dialog_spec.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+# This file is part of Alexandria.
+#
+# See the file README.md for authorship and licensing information.
+
+require File.dirname(__FILE__) + '/../../spec_helper'
+
+describe Alexandria::UI::BookPropertiesDialog do
+  it 'works' do
+    parent = Gtk::Window.new :toplevel
+    library = instance_double(Alexandria::Library, name: 'Bar Library', cover: '')
+    book = Alexandria::Book.new('Foo Book', ['Jane Doe'], '98765432', 'Bar Publisher',
+                                1972, 'edition')
+    described_class.new parent, library, book
+  end
+end
